# openapi.yaml
openapi: 3.0.3
info:
  title: Yape Code Challenge API
  version: "1.0.0"
  description: |
    Transaction API with Anti-Fraud (event-driven). 
    - Create transaction (starts as `pending` and is later updated to `approved`/`rejected` by anti-fraud)
    - Get transaction by id
servers:
  - url: http://localhost:3000
    description: Local Docker
tags:
  - name: Health
  - name: Transactions
paths:
  /api/v1/health:
    get:
      tags: [Health]
      summary: Health check
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
              examples:
                ok:
                  value: { status: "ok" }
  /api/v1/transactions:
    post:
      tags: [Transactions]
      summary: Create a new transaction
      description: |
        Creates a transaction with status `pending`. The Anti-Fraud service consumes the event
        and publishes a decision to update the status to `approved` or `rejected`.
      operationId: createTransaction
      parameters:
        - in: header
          name: Idempotency-Key
          required: false
          description: Unique key to make the request idempotent
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionRequest"
            examples:
              approved:
                value:
                  accountExternalIdDebit: "11111111-1111-4111-8111-111111111111"
                  accountExternalIdCredit: "22222222-2222-4222-8222-222222222222"
                  tranferTypeId: 1
                  value: 500
              rejected:
                value:
                  accountExternalIdDebit: "11111111-1111-4111-8111-111111111111"
                  accountExternalIdCredit: "22222222-2222-4222-8222-222222222222"
                  tranferTypeId: 1
                  value: 1500
      responses:
        "201":
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateTransactionResponse"
              examples:
                created:
                  value: { transactionExternalId: "b4f9c2e7a1d3..." }
        "400":
          description: Validation error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "422":
          description: Domain rule violation (e.g., same debit/credit account)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /api/v1/transactions/{transactionExternalId}:
    get:
      tags: [Transactions]
      summary: Get transaction by id
      operationId: getTransaction
      parameters:
        - in: path
          name: transactionExternalId
          required: true
          schema:
            type: string
          description: Transaction external identifier
      responses:
        "200":
          description: Transaction found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
              examples:
                approved:
                  value:
                    transactionExternalId: "b4f9c2e7a1d3..."
                    transactionType: { name: "1" }
                    transactionStatus: { name: "approved" }
                    value: 500
                    createdAt: "2025-09-19T22:15:10.123Z"
                rejected:
                  value:
                    transactionExternalId: "c7e2d1a9f0..."
                    transactionType: { name: "1" }
                    transactionStatus: { name: "rejected" }
                    value: 1500
                    createdAt: "2025-09-19T22:15:10.123Z"
                pending:
                  value:
                    transactionExternalId: "a1b2c3d4e5..."
                    transactionType: { name: "1" }
                    transactionStatus: { name: "pending" }
                    value: 120
                    createdAt: "2025-09-19T22:15:10.123Z"
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
components:
  schemas:
    HealthResponse:
      type: object
      properties:
        status:
          type: string
          example: ok
    CreateTransactionRequest:
      type: object
      required: [accountExternalIdDebit, accountExternalIdCredit, tranferTypeId, value]
      properties:
        accountExternalIdDebit:
          type: string
          format: uuid
          description: UUID v4 of debit account
        accountExternalIdCredit:
          type: string
          format: uuid
          description: UUID v4 of credit account
        tranferTypeId:
          description: Transfer type ID (1|2|3)
          oneOf:
            - type: integer
              enum: [1, 2, 3]
            - type: string
              enum: ["1", "2", "3"]
        value:
          type: integer
          minimum: 1
          maximum: 1000000
          description: Amount in integer cents (1..1,000,000)
    CreateTransactionResponse:
      type: object
      properties:
        transactionExternalId:
          type: string
          description: Newly created transaction external ID
    Transaction:
      type: object
      properties:
        transactionExternalId:
          type: string
        transactionType:
          $ref: "#/components/schemas/TransactionType"
        transactionStatus:
          $ref: "#/components/schemas/TransactionStatus"
        value:
          type: integer
          description: Amount in integer cents
        createdAt:
          type: string
          format: date-time
      required: [transactionExternalId, transactionType, transactionStatus, value, createdAt]
    TransactionType:
      type: object
      properties:
        name:
          type: string
          description: Transfer type (stringified id)
          example: "1"
    TransactionStatus:
      type: object
      properties:
        name:
          type: string
          enum: [pending, approved, rejected]
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          example: https://zod.dev/validation-error
        title:
          type: string
          example: Validation Error
        status:
          type: integer
          example: 400
        detail:
          description: Additional details (array of issues for 400, message for domain rules)